requirements.txt
twitchio>=2.2.0,<=2.3.0
python-dotenv>=0.20.0,<=0.21.0